+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/variables/IBuildPropertiesInitializer.java
@@ -1,7 +1,0 @@
-package org.objectstyle.wolips.variables;
-
-public interface IBuildPropertiesInitializer {
-	public void initializeDefaults(BuildProperties buildProperties);
-	
-	public void initialize(BuildProperties buildProperties);
-}